API related functionality in a separate module
