+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/arbitrum_controller.ex
@@ -12,9 +12,11 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   import Explorer.Chain.Arbitrum.DaMultiPurposeRecord.Helper, only: [calculate_celestia_data_key: 2]
 
   alias Explorer.Arbitrum.ClaimRollupMessage
-  alias Explorer.Chain.Arbitrum.{L1Batch, Message, Reader}
+  alias Explorer.Chain.Arbitrum.{L1Batch, Message}
   alias Explorer.Chain.Hash
   alias Explorer.PagingOptions
+
+  alias Explorer.Chain.Arbitrum.Reader.API, as: Reader
 
   action_fallback(BlockScoutWeb.API.V2.FallbackController)
 
@@ -28,7 +30,6 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
     options =
       params
       |> paging_options()
-      |> Keyword.put(:api?, true)
 
     {messages, next_page} =
       direction
@@ -58,7 +59,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   def messages_count(conn, %{"direction" => direction} = _params) do
     conn
     |> put_status(200)
-    |> render(:arbitrum_messages_count, %{count: Reader.messages_count(direction, api?: true)})
+    |> render(:arbitrum_messages_count, %{count: Reader.messages_count(direction)})
   end
 
   @doc """
@@ -124,11 +125,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   """
   @spec batch(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def batch(conn, %{"batch_number" => batch_number} = _params) do
-    case Reader.batch(
-           batch_number,
-           necessity_by_association: @batch_necessity_by_association,
-           api?: true
-         ) do
+    case Reader.batch(batch_number, necessity_by_association: @batch_necessity_by_association) do
       {:ok, batch} ->
         conn
         |> put_status(200)
@@ -146,7 +143,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   @spec batch_by_data_availability_info(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def batch_by_data_availability_info(conn, %{"data_hash" => data_hash} = _params) do
     # In case of AnyTrust, `data_key` is the hash of the data itself
-    case Reader.get_da_record_by_data_key(data_hash, api?: true) do
+    case Reader.get_da_record_by_data_key(data_hash) do
       {:ok, {batch_number, _}} ->
         batch(conn, %{"batch_number" => batch_number})
 
@@ -162,7 +159,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
     # In case of Celestia, `data_key` is the hash of the height and the commitment hash
     with {:ok, :hash, transaction_commitment_hash} <- parse_block_hash_or_number_param(transaction_commitment),
          key <- calculate_celestia_data_key(height, transaction_commitment_hash) do
-      case Reader.get_da_record_by_data_key(key, api?: true) do
+      case Reader.get_da_record_by_data_key(key) do
         {:ok, {batch_number, _}} ->
           batch(conn, %{"batch_number" => batch_number})
 
@@ -182,7 +179,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   def batches_count(conn, _params) do
     conn
     |> put_status(200)
-    |> render(:arbitrum_batches_count, %{count: Reader.batches_count(api?: true)})
+    |> render(:arbitrum_batches_count, %{count: Reader.batches_count()})
   end
 
   @doc """
@@ -194,7 +191,6 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
       params
       |> paging_options()
       |> Keyword.put(:necessity_by_association, @batch_necessity_by_association)
-      |> Keyword.put(:api?, true)
       |> Reader.batches()
       |> split_list_by_page()
 
@@ -222,7 +218,6 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
     batches =
       []
       |> Keyword.put(:necessity_by_association, @batch_necessity_by_association)
-      |> Keyword.put(:api?, true)
       |> Keyword.put(:committed?, true)
       |> Reader.batches()
 
@@ -242,7 +237,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   end
 
   defp batch_latest_number do
-    case Reader.batch(:latest, api?: true) do
+    case Reader.batch(:latest) do
       {:ok, batch} -> batch.number
       {:error, :not_found} -> 0
     end
@@ -253,7 +248,7 @@ defmodule BlockScoutWeb.API.V2.ArbitrumController do
   """
   @spec recent_messages_to_l2(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def recent_messages_to_l2(conn, _params) do
-    messages = Reader.relayed_l1_to_l2_messages(paging_options: %PagingOptions{page_size: 6}, api?: true)
+    messages = Reader.relayed_l1_to_l2_messages(paging_options: %PagingOptions{page_size: 6})
 
     conn
     |> put_status(200)
